requirements.txt
anthropic>=0.86.0
litellm>=1.80.0
aurite>=0.4.2
beautifulsoup4>=4.14.0
curl_cffi>=0.15.0
fastapi>=0.115.0
httpx>=0.28.0
lxml>=6.0.0
mcp>=1.26.0
pydantic>=2.12.0
pypdf>=6.0.0
python-dotenv>=1.2.0
python-multipart>=0.0.22
requests>=2.33.0
uvicorn>=0.35.0
